Redo Sound System
Sounds are now "requested" via increasing a u16 variable in the SFXResource
Then a systems in the sound module will add appropriate sounds to a queue and then run through it until it's empty and all requests are fulfilled.
There is also a max bounce sound constant as well

It seems like it's doubling sounds sometimes, and maybe playing them SLIGHTLY late other times. But so far there have been no buggy sounding issues like there were before with a lot of enemies.